Implement logic to fetch the report count of a group

+++ SmartAlertAPI/src/main/java/com/unipi/smartalert/dtos/ReportGroupDTO.java
@@ -16,7 +16,7 @@ public class ReportGroupDTO {
     private long groupId;
     private Map<String, String> incidentCategory;
     private String lastReportTimestamp;
-    private int reportCount;
+    private long reportCount;
     private LocationDTO location;
     private double searchRadiusInMeters;
